added feedback and percentage
I added code to implement a test score and feedback for how you did. At this point the code is a little messy, as I wasn't able to implement the variables to shorten the if statements at the end without breaking it. I have to move on to the CS20 projects before I can commit to this due to the fact that this quarter ends in about a week and a half.